tests: index: move fixtures to conftest
